Renames `toParts` to `msToParts`
